requirements.txt
python-binance
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
ta
torch
jupyter
